+++ ProjetoBAN2Fase2/src/infrastructure2/ProdutosRepository.java
@@ -3,22 +3,26 @@ package infrastructure2;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Updates;
+import com.mongodb.client.model.UpdateOptions;
 import domain2.Categoria;
 import domain2.Fornecedor;
 import domain2.Produto;
+import domain2.Venda;
 
 import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.Sorts;
 import org.bson.conversions.Bson;
 
 import java.util.ArrayList;
-import java.util.logging.Filter;
+import java.util.List;
 
 public class ProdutosRepository {
     private final MongoCollection<Produto> produtosCollection;
+    private final MongoCollection<Venda> vendasCollection;
 
     public ProdutosRepository(MongoDatabase database) {
         produtosCollection = database.getCollection("produtos", Produto.class);
+        vendasCollection = database.getCollection("vendas", Venda.class);
     }
 
     public void createProduto(Produto produto) {
@@ -26,6 +30,12 @@ public class ProdutosRepository {
     }
 
     public void deleteProduto(int id) {
+        long count = vendasCollection.countDocuments(Filters.eq("vendaProdutos.produto._id", id));
+
+        if (count > 0) {
+            throw new IllegalStateException("ERRO: Não é possível excluir este produto pois ele já foi vendido e consta em registros de vendas.");
+        }
+
         Bson filter = Filters.eq("_id", id);
         produtosCollection.deleteOne(filter);
     }
@@ -42,6 +52,15 @@ public class ProdutosRepository {
     public void updateProduto(Produto produto) {
         Bson filter = Filters.eq("_id", produto.getId());
         produtosCollection.replaceOne(filter, produto);
+
+        Bson filterVendas = Filters.eq("vendaProdutos.produto._id", produto.getId());
+        Bson updateVebdas = Updates.set("vendaProdutos.$[elem].produto", produto);
+
+        UpdateOptions options = new UpdateOptions().arrayFilters(
+                List.of(Filters.eq("elem.produto._id", produto.getId()))
+        );
+
+        vendasCollection.updateMany(filterVendas,updateVebdas,options);
     }
 
     public void decrementarQuantidade(int id, int quantidade) {
